lazycat-appstore-analyzer/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
pandas==2.0.3
requests==2.31.0
beautifulsoup4==4.12.2
authlib==1.2.1
python-dotenv==1.0.0